Paragraph on the importance of set -e
This is something I had to learn the hard way after writing way to many
scripts without set -e which eventually turned into a nightmare to
debug. Everything can fail, even a simple mktemp (because the /tmp
filesystem is full) can fail, and if it does, you don't want your
program flow to continue like nothing happened. It's especially useful
in conjunction with a cleanup handler (trap).

Another paragraph on pipefail would be great as well:
http://www.gnu.org/software/bash/manual/html_node/Pipelines.html

Signed-off-by: Valentin Dornauer <valentin@unimplemented.org>