Paragraph on set -u
'set -u' has in my experience proven to be very useful in the
development of new scripts, since it uncovers typos in variable names
and avoids function calls which are missing parameters. It's mandatory
for shell libraries because it will otherwise break programs that use
that library and (rightfully!) chose to set -u.

Signed-off-by: Valentin Dornauer <valentin@unimplemented.org>